gen_advc_tasks/requirements.txt
python-dotenv
openai
tqdm
pillow